Port strict-export-declare-modifiers
