+++ packages/eslint-plugin/test/__file_snapshots__/types/strict-export-declare-modifiers/bad4.d.ts.lint
@@ -1,0 +1,12 @@
+types/strict-export-declare-modifiers/bad4.d.ts
+  2:3  error  'export' keyword is redundant here because all declarations in this module are exported automatically. If you have a good reason to export some declarations and not others, add 'export {}' to the module to shut off automatic exporting  @definitelytyped/strict-export-declare-modifiers
+
+✖ 1 problem (1 error, 0 warnings)
+
+==== types/strict-export-declare-modifiers/bad4.d.ts ====
+
+    export namespace M {
+      export function f(): void;
+      ~~~~~~
+!!! @definitelytyped/strict-export-declare-modifiers: 'export' keyword is redundant here because all declarations in this module are exported automatically. If you have a good reason to export some declarations and not others, add 'export {}' to the module to shut off automatic exporting.
+    }
