+++ packages/eslint-plugin/test/__file_snapshots__/types/strict-export-declare-modifiers/bad5.d.ts.lint
@@ -1,0 +1,11 @@
+types/strict-export-declare-modifiers/bad5.d.ts
+  1:11  error  All declarations in this module are exported automatically. Prefer to explicitly write 'export' for clarity. If you have a good reason not to export this declaration, add 'export {}' to the module to shut off automatic exporting  @definitelytyped/strict-export-declare-modifiers
+
+✖ 1 problem (1 error, 0 warnings)
+
+==== types/strict-export-declare-modifiers/bad5.d.ts ====
+
+    interface I {}
+              ~
+!!! @definitelytyped/strict-export-declare-modifiers: All declarations in this module are exported automatically. Prefer to explicitly write 'export' for clarity. If you have a good reason not to export this declaration, add 'export {}' to the module to shut off automatic exporting.
+    export namespace M {}
