+++ packages/eslint-plugin/test/__file_snapshots__/types/strict-export-declare-modifiers/good2.d.ts.lint
@@ -1,0 +1,5 @@
+No errors
+
+==== types/strict-export-declare-modifiers/good2.d.ts ====
+
+    export function f() {}
