+++ packages/eslint-plugin/test/__file_snapshots__/types/strict-export-declare-modifiers/good8.d.ts.lint
@@ -1,0 +1,9 @@
+No errors
+
+==== types/strict-export-declare-modifiers/good8.d.ts ====
+
+    import * as foo from "foo";
+    import foo = require("foo");
+    export { foo };
+    export { foo } from "foo";
+    export as namespace Foo;
