+++ packages/eslint-plugin/test/__file_snapshots__/types/strict-export-declare-modifiers/strict-export-declare-modifiers-tests.ts.lint
@@ -1,0 +1,10 @@
+types/strict-export-declare-modifiers/strict-export-declare-modifiers-tests.ts
+  1:1  error  File has no content  @definitelytyped/no-useless-files
+
+✖ 1 problem (1 error, 0 warnings)
+
+==== types/strict-export-declare-modifiers/strict-export-declare-modifiers-tests.ts ====
+
+    
+    ~
+!!! @definitelytyped/no-useless-files: File has no content.
