+++ packages/eslint-plugin/test/fixtures/types/strict-export-declare-modifiers/bad1.d.ts
@@ -1,0 +1,1 @@
+declare interface I {}
