+++ packages/eslint-plugin/test/fixtures/types/strict-export-declare-modifiers/bad2.d.ts
@@ -1,0 +1,3 @@
+declare namespace M {
+    export const x: number;
+}
