+++ packages/eslint-plugin/test/fixtures/types/strict-export-declare-modifiers/bad3.d.ts
@@ -1,0 +1,1 @@
+export declare function f(): void;
