+++ packages/eslint-plugin/test/fixtures/types/strict-export-declare-modifiers/bad4.d.ts
@@ -1,0 +1,3 @@
+export namespace M {
+  export function f(): void;
+}
