+++ packages/eslint-plugin/test/fixtures/types/strict-export-declare-modifiers/bad5.d.ts
@@ -1,0 +1,2 @@
+interface I {}
+export namespace M {}
