+++ packages/eslint-plugin/test/fixtures/types/strict-export-declare-modifiers/bad6.d.ts
@@ -1,0 +1,2 @@
+declare function g(): void;
+export namespace M {}
