+++ packages/eslint-plugin/test/fixtures/types/strict-export-declare-modifiers/bad7.d.ts
@@ -1,0 +1,2 @@
+declare namespace N {}
+export namespace M {}
