+++ packages/eslint-plugin/test/fixtures/types/strict-export-declare-modifiers/good1.ts
@@ -1,0 +1,1 @@
+export declare class C {}
