+++ packages/eslint-plugin/test/fixtures/types/strict-export-declare-modifiers/good2.d.ts
@@ -1,0 +1,1 @@
+export function f() {}
