+++ packages/eslint-plugin/test/fixtures/types/strict-export-declare-modifiers/good4.d.ts
@@ -1,0 +1,1 @@
+declare namespace N {}
