+++ packages/eslint-plugin/test/fixtures/types/strict-export-declare-modifiers/good5.d.ts
@@ -1,0 +1,1 @@
+interface J {}
