+++ packages/eslint-plugin/test/fixtures/types/strict-export-declare-modifiers/good6.ts
@@ -1,0 +1,3 @@
+namespace N {
+    export const x: number;
+}
