+++ packages/eslint-plugin/test/fixtures/types/strict-export-declare-modifiers/good7.d.ts
@@ -1,0 +1,2 @@
+declare class Foo2 {}
+export { Foo2 as Bar2 }
