+++ packages/eslint-plugin/test/fixtures/types/strict-export-declare-modifiers/good8.d.ts
@@ -1,0 +1,5 @@
+import * as foo from "foo";
+import foo = require("foo");
+export { foo };
+export { foo } from "foo";
+export as namespace Foo;
