+++ packages/eslint-plugin/test/fixtures/types/strict-export-declare-modifiers/good9.d.ts
@@ -1,0 +1,3 @@
+import * as bar from "bar";
+import bar = require("bar");
+export as namespace Bar;
