+++ packages/eslint-plugin/test/fixtures/types/strict-export-declare-modifiers/package.json
@@ -1,0 +1,8 @@
+{
+    "name": "@types/strict-export-declare-modifiers",
+    "version": "2.0.9999",
+    "owners": [],
+    "devDependencies": {
+        "@types/strict-export-declare-modifiers": "workspace:."
+    }
+}
